+++ script/tool/lib/src/native_test_command.dart
@@ -35,6 +35,7 @@ class NativeTestCommand extends PackageLoopingCommand {
           'See https://developer.apple.com/library/archive/technotes/tn2339/_index.html#//apple_ref/doc/uid/DTS40014588-CH1-UNIT '
           'for details on how to specify the destination.',
     );
+    argParser.addFlag(kPlatformAndroid, help: 'Runs Android tests');
     argParser.addFlag(kPlatformIos, help: 'Runs iOS tests');
     argParser.addFlag(kPlatformMacos, help: 'Runs macOS tests');
 
@@ -45,6 +46,8 @@ class NativeTestCommand extends PackageLoopingCommand {
     argParser.addFlag(_integrationTestFlag,
         help: 'Runs native integration (UI) tests', defaultsTo: true);
   }
+
+  static const String _gradleWrapper = 'gradlew';
 
   // The device destination flags for iOS tests.
   List<String> _iosDestinationFlags = <String>[];
@@ -59,6 +62,7 @@ class NativeTestCommand extends PackageLoopingCommand {
 Runs native unit tests and native integration tests.
 
 Currently supported platforms:
+- Android (unit tests only)
 - iOS: requires 'xcrun' to be in your path.
 - macOS: requires 'xcrun' to be in your path.
 
@@ -73,6 +77,7 @@ this command.
   @override
   Future<void> initializeRun() async {
     _platforms = <String, _PlatformDetails>{
+      kPlatformAndroid: _PlatformDetails('Android', _testAndroid),
       kPlatformIos: _PlatformDetails('iOS', _testIos),
       kPlatformMacos: _PlatformDetails('macOS', _testMacOs),
     };
@@ -89,6 +94,11 @@ this command.
     if (!(getBoolArg(_unitTestFlag) || getBoolArg(_integrationTestFlag))) {
       printError('At least one test type must be enabled.');
       throw ToolExit(exitInvalidArguments);
+    }
+
+    if (getBoolArg(kPlatformAndroid) && getBoolArg(_integrationTestFlag)) {
+      logWarning('This command currently only supports unit tests for Android. '
+          'See https://github.com/flutter/flutter/issues/86490.');
     }
 
     // iOS-specific run-level state.
@@ -131,34 +141,105 @@ this command.
       integration: getBoolArg(_integrationTestFlag),
     );
 
-    final List<String> failures = <String>[];
     bool ranTests = false;
+    bool failed = false;
+    final List<String> failureMessages = <String>[];
     for (final String platform in testPlatforms) {
       final _PlatformDetails platformInfo = _platforms[platform]!;
-      final RunState result = await platformInfo.testFunction(package, mode);
-      ranTests |= result != RunState.skipped;
-      if (result == RunState.failed) {
-        failures.add(platformInfo.label);
+      print('Running tests for ${platformInfo.label}...');
+      print('----------------------------------------');
+      final _PlatformResult result =
+          await platformInfo.testFunction(package, mode);
+      ranTests |= result.state != RunState.skipped;
+      if (result.state == RunState.failed) {
+        failed = true;
+
+        final String? error = result.error;
+        // Only provide the failing platforms in the failure details if testing
+        // multiple platforms, otherwise it's just noise.
+        if (_requestedPlatforms.length > 1) {
+          failureMessages.add(error != null
+              ? '${platformInfo.label}: $error'
+              : platformInfo.label);
+        } else if (error != null) {
+          // If there's only one platform, only provide error details in the
+          // summary if the platform returned a message.
+          failureMessages.add(error);
+        }
       }
     }
 
     if (!ranTests) {
       return PackageResult.skip('No tests found.');
     }
-    // Only provide the failing platforms in the failure details if testing
-    // multiple platforms, otherwise it's just noise.
-    return failures.isEmpty
-        ? PackageResult.success()
-        : PackageResult.fail(
-            _requestedPlatforms.length > 1 ? failures : <String>[]);
+    return failed
+        ? PackageResult.fail(failureMessages)
+        : PackageResult.success();
   }
 
-  Future<RunState> _testIos(Directory plugin, _TestMode mode) {
+  Future<_PlatformResult> _testAndroid(Directory plugin, _TestMode mode) async {
+    final List<Directory> examplesWithTests = <Directory>[];
+    for (final Directory example in getExamplesForPlugin(plugin)) {
+      if (!isFlutterPackage(example)) {
+        continue;
+      }
+      if (example
+              .childDirectory('android')
+              .childDirectory('app')
+              .childDirectory('src')
+              .childDirectory('test')
+              .existsSync() ||
+          example.parent
+              .childDirectory('android')
+              .childDirectory('src')
+              .childDirectory('test')
+              .existsSync()) {
+        examplesWithTests.add(example);
+      } else {
+        _printNoExampleTestsMessage(example, 'Android');
+      }
+    }
+
+    if (examplesWithTests.isEmpty) {
+      return _PlatformResult(RunState.skipped);
+    }
+
+    bool failed = false;
+    bool hasMissingBuild = false;
+    for (final Directory example in examplesWithTests) {
+      final String exampleName = getPackageDescription(example);
+      _printRunningExampleTestsMessage(example, 'Android');
+
+      final Directory androidDirectory = example.childDirectory('android');
+      final File gradleFile = androidDirectory.childFile(_gradleWrapper);
+      if (!gradleFile.existsSync()) {
+        printError('ERROR: Run "flutter build apk" on $exampleName, or run '
+            'this tool\'s "build-examples --apk" command, '
+            'before executing tests.');
+        failed = true;
+        hasMissingBuild = true;
+        continue;
+      }
+
+      final int exitCode = await processRunner.runAndStream(
+          gradleFile.path, <String>['testDebugUnitTest', '--info'],
+          workingDir: androidDirectory);
+      if (exitCode != 0) {
+        printError('$exampleName tests failed.');
+        failed = true;
+      }
+    }
+    return _PlatformResult(failed ? RunState.failed : RunState.succeeded,
+        error:
+            hasMissingBuild ? 'Examples must be built before testing.' : null);
+  }
+
+  Future<_PlatformResult> _testIos(Directory plugin, _TestMode mode) {
     return _runXcodeTests(plugin, 'iOS', mode,
         extraFlags: _iosDestinationFlags);
   }
 
-  Future<RunState> _testMacOs(Directory plugin, _TestMode mode) {
+  Future<_PlatformResult> _testMacOs(Directory plugin, _TestMode mode) {
     return _runXcodeTests(plugin, 'macOS', mode);
   }
 
@@ -167,7 +248,7 @@ this command.
   ///
   /// The tests targets must be added to the Xcode project of the example app,
   /// usually at "example/{ios,macos}/Runner.xcworkspace".
-  Future<RunState> _runXcodeTests(
+  Future<_PlatformResult> _runXcodeTests(
     Directory plugin,
     String platform,
     _TestMode mode, {
@@ -183,8 +264,7 @@ this command.
     // Assume skipped until at least one test has run.
     RunState overallResult = RunState.skipped;
     for (final Directory example in getExamplesForPlugin(plugin)) {
-      final String examplePath =
-          getRelativePosixPath(example, from: plugin.parent);
+      final String exampleName = getPackageDescription(example);
 
       if (testTarget != null) {
         final Directory project = example
@@ -193,16 +273,16 @@ this command.
         final bool? hasTarget =
             await _xcode.projectHasTarget(project, testTarget);
         if (hasTarget == null) {
-          printError('Unable to check targets for $examplePath.');
+          printError('Unable to check targets for $exampleName.');
           overallResult = RunState.failed;
           continue;
         } else if (!hasTarget) {
-          print('No "$testTarget" target in $examplePath; skipping.');
+          print('No "$testTarget" target in $exampleName; skipping.');
           continue;
         }
       }
 
-      print('Running $platform tests for $examplePath...');
+      _printRunningExampleTestsMessage(example, platform);
       final int exitCode = await _xcode.runXcodeBuild(
         example,
         actions: <String>['test'],
@@ -220,10 +300,10 @@ this command.
       const int _xcodebuildNoTestExitCode = 66;
       switch (exitCode) {
         case _xcodebuildNoTestExitCode:
-          print('No tests found for $examplePath');
+          _printNoExampleTestsMessage(example, platform);
           continue;
         case 0:
-          printSuccess('Successfully ran $platform xctest for $examplePath');
+          printSuccess('Successfully ran $platform xctest for $exampleName');
           // If this is the first test, assume success until something fails.
           if (overallResult == RunState.skipped) {
             overallResult = RunState.succeeded;
@@ -235,13 +315,25 @@ this command.
           break;
       }
     }
-    return overallResult;
+    return _PlatformResult(overallResult);
+  }
+
+  /// Prints a standard format message indicating that [platform] tests for
+  /// [plugin]'s [example] are about to be run.
+  void _printRunningExampleTestsMessage(Directory example, String platform) {
+    print('Running $platform tests for ${getPackageDescription(example)}...');
+  }
+
+  /// Prints a standard format message indicating that no tests were found for
+  /// [plugin]'s [example] for [platform].
+  void _printNoExampleTestsMessage(Directory example, String platform) {
+    print('No $platform tests found for ${getPackageDescription(example)}');
   }
 }
 
 // The type for a function that takes a plugin directory and runs its native
 // tests for a specific platform.
-typedef _TestFunction = Future<RunState> Function(Directory, _TestMode);
+typedef _TestFunction = Future<_PlatformResult> Function(Directory, _TestMode);
 
 /// A collection of information related to a specific platform.
 class _PlatformDetails {
@@ -266,4 +358,20 @@ class _TestMode {
 
   bool get integrationOnly => integration && !unit;
   bool get unitOnly => unit && !integration;
+}
+
+/// The result of running a single platform's tests.
+class _PlatformResult {
+  _PlatformResult(this.state, {this.error});
+
+  /// The overall state of the platform's tests. This should be:
+  /// - failed if any tests failed.
+  /// - succeeded if at least one test ran, and all tests passed.
+  /// - skipped if no tests ran.
+  final RunState state;
+
+  /// An optional error string to include in the summary for this platform.
+  ///
+  /// Ignored unless [state] is `failed`.
+  final String? error;
 }